refactor: port, adapter usecase 단위로 패키지 구조 수정 및 rename
