+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/body_type/usecase/get_body_types/GetBodyTypesPort.java
@@ -1,4 +1,4 @@
-package softeer.be_my_car_master.api.body_type.usecase.port;
+package softeer.be_my_car_master.api.body_type.usecase.get_body_types;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import softeer.be_my_car_master.domain.body_type.BodyType;
 import softeer.be_my_car_master.global.annotation.Port;
 
 @Port
-public interface BodyTypePort {
+public interface GetBodyTypesPort {
 
 	List<BodyType> findSelectableBodyTypesByModelId(Long modelId);
 }
