+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/body_type/usecase/get_body_types/GetBodyTypesUseCase.java
@@ -1,10 +1,9 @@
-package softeer.be_my_car_master.api.body_type.usecase;
+package softeer.be_my_car_master.api.body_type.usecase.get_body_types;
 
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
 import softeer.be_my_car_master.api.body_type.dto.response.GetBodyTypesResponse;
-import softeer.be_my_car_master.api.body_type.usecase.port.BodyTypePort;
 import softeer.be_my_car_master.domain.body_type.BodyType;
 import softeer.be_my_car_master.global.annotation.UseCase;
 
@@ -12,10 +11,10 @@ import softeer.be_my_car_master.global.annotation.UseCase;
 @RequiredArgsConstructor
 public class GetBodyTypesUseCase {
 
-	private final BodyTypePort bodyTypePort;
+	private final GetBodyTypesPort getBodyTypePort;
 
 	public GetBodyTypesResponse execute(Long modelId) {
-		List<BodyType> selectableBodyTypes = bodyTypePort.findSelectableBodyTypesByModelId(modelId);
+		List<BodyType> selectableBodyTypes = getBodyTypePort.findSelectableBodyTypesByModelId(modelId);
 		return GetBodyTypesResponse.from(selectableBodyTypes);
 	}
 }
