+++ BE-MyCarMaster/src/test/java/softeer/be_my_car_master/api/body_type/usecase/GetBodyTypesUseCaseTest.java
@@ -15,7 +15,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import softeer.be_my_car_master.api.body_type.dto.response.BodyTypeDto;
 import softeer.be_my_car_master.api.body_type.dto.response.GetBodyTypesResponse;
-import softeer.be_my_car_master.api.body_type.usecase.port.BodyTypePort;
+import softeer.be_my_car_master.api.body_type.usecase.get_body_types.GetBodyTypesPort;
+import softeer.be_my_car_master.api.body_type.usecase.get_body_types.GetBodyTypesUseCase;
 import softeer.be_my_car_master.domain.body_type.BodyType;
 
 @ExtendWith(MockitoExtension.class)
@@ -26,7 +27,7 @@ class GetBodyTypesUseCaseTest {
 	private GetBodyTypesUseCase getBodyTypesUseCase;
 
 	@Mock
-	private BodyTypePort bodyTypePort;
+	private GetBodyTypesPort getBodyTypePort;
 
 	@Test
 	@DisplayName("바디타입 목록을 조회합니다")
@@ -40,7 +41,7 @@ class GetBodyTypesUseCaseTest {
 			.ratio(22)
 			.imgUrl("imgUrl")
 			.build();
-		given(bodyTypePort.findSelectableBodyTypesByModelId(any())).willReturn(Arrays.asList(bodyType));
+		given(getBodyTypePort.findSelectableBodyTypesByModelId(any())).willReturn(Arrays.asList(bodyType));
 
 		// when
 		GetBodyTypesResponse getBodyTypesResponse = getBodyTypesUseCase.execute(1L);
